remove subsystem file
